refactor build fn to use a class for managing state
