+++ lib/src/generate/exceptions.dart
@@ -1,0 +1,11 @@
+// Copyright (c) 2016, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+class ConcurrentBuildException implements Exception {
+
+  const ConcurrentBuildException();
+
+  @override
+  String toString() =>
+      'ConcurrentBuildException: Only one build may be running at a time.';
+}
